+++ CHANGELOG.md
@@ -29,6 +29,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## [Unreleased]
+
+### Bug Fixes
+
+* [#39](https://github.com/JulianToledano/goingecko/pull/39) Change CodeStats types from int16 to int64.
+
 ## [v3.0.0](https://github.com/JulianToledano/goingecko/releases/tag/v3.0.0) - 2025.01-25
 
 ### Improvements